(By Sklower) checkpoint the current state of Cherenson's work.
SCCS-vsn: sys/netinet/in_pcb.c 7.24
SCCS-vsn: sys/netinet/ip_icmp.h 7.6
SCCS-vsn: sys/netinet/if_ether.h 7.9
SCCS-vsn: sys/netinet/ip_output.c 7.27
SCCS-vsn: sys/netinet/in.c 7.27
SCCS-vsn: sys/netinet/in_proto.c 7.9
SCCS-vsn: sys/netinet/tcp_usrreq.c 7.18
SCCS-vsn: sys/netinet/tcp_input.c 7.30
SCCS-vsn: sys/netinet/tcp_timer.c 7.20
SCCS-vsn: sys/netinet/ip_icmp.c 7.19
SCCS-vsn: sys/netinet/ip_var.h 7.9
SCCS-vsn: sys/netinet/tcp_subr.c 7.25
SCCS-vsn: sys/netinet/icmp_var.h 7.6
SCCS-vsn: sys/netinet/udp_usrreq.c 7.27
SCCS-vsn: sys/netinet/tcp.h 7.8
SCCS-vsn: sys/netinet/tcp_var.h 7.11
SCCS-vsn: sys/netinet/udp_var.h 7.8
SCCS-vsn: sys/netinet/igmp.c 7.3
SCCS-vsn: sys/netinet/ip_input.c 7.24
SCCS-vsn: sys/netinet/raw_ip.c 7.11
SCCS-vsn: sys/netinet/tcp_output.c 7.25
SCCS-vsn: sys/netinet/if_ether.c 7.27
SCCS-vsn: sys/netinet/igmp_var.h 7.2
SCCS-vsn: sys/netinet/ip_mroute.c 7.5

+++ usr/src/sys/netinet/udp_var.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)udp_var.h	7.7 (Berkeley) %G%
+ *	@(#)udp_var.h	7.8 (Berkeley) %G%
  */
 
 /*
@@ -28,16 +28,16 @@ struct	udpiphdr {
 
 struct	udpstat {
 				/* input statistics: */
-	int	udps_ipackets;		/* total input packets */
-	int	udps_hdrops;		/* packet shorter than header */
-	int	udps_badsum;		/* checksum error */
-	int	udps_badlen;		/* data length larger than packet */
-	int	udps_noport;		/* no socket on port */
-	int	udps_noportbcast;	/* of above, arrived as broadcast */
-	int	udps_fullsock;		/* not delivered, input socket full */
-	int	udpps_pcbcachemiss;	/* input packets missing pcb cache */
+	u_long	udps_ipackets;		/* total input packets */
+	u_long	udps_hdrops;		/* packet shorter than header */
+	u_long	udps_badsum;		/* checksum error */
+	u_long	udps_badlen;		/* data length larger than packet */
+	u_long	udps_noport;		/* no socket on port */
+	u_long	udps_noportbcast;	/* of above, arrived as broadcast */
+	u_long	udps_fullsock;		/* not delivered, input socket full */
+	u_long	udpps_pcbcachemiss;	/* input packets missing pcb cache */
 				/* output statistics: */
-	int	udps_opackets;		/* total output packets */
+	u_long	udps_opackets;		/* total output packets */
 };
 
 #define	UDP_TTL		30	/* default time to live for UDP packets */
